+++ RPGAapp/Data Classes/Item+CoreDataProperties.swift
@@ -21,7 +21,7 @@ extension Item {
     @NSManaged public var measure: String?
     @NSManaged public var name: String?
     @NSManaged public var price: Double
-    @NSManaged public var propability: Int16
+    @NSManaged public var propability: Int64
     @NSManaged public var quantity: Int16
     @NSManaged public var rarity: Int16
     @NSManaged public var category: Category?
